dev/4_microservices/search/searchUsername/requirements.txt
grpc-interceptor == 0.11.0
grpcio-tools == 1.44.0
google-cloud-bigquery == 3.0.1